requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
